logical device done

+++ src/renderer/queue_family_indices.rs
@@ -1,6 +1,6 @@
 #[derive(Default, Clone)]
 pub struct QueueFamilyIndices {
-  pub graphics_queue_family: Option<usize>,
+  pub graphics_queue_family: Option<u32>,
 }
 impl QueueFamilyIndices {
   pub fn is_complete(&self) -> bool {
